Refactor database collection slightly
