+++ server/database.ts
@@ -1,14 +1,6 @@
+import { SqLiteDatabaseCollection } from '#server/database/DatabaseCollection';
 import { Booking } from '#shared/models';
-import { DatabaseSync, StatementSync, type SQLInputValue } from 'node:sqlite';
-import z, { ZodObject } from 'zod';
-
-interface DatabaseCollection<T> {
-  getAll(): T[];
-  getSingle(id: number): T;
-  update(entry: T): void;
-  insert(entry: T): void;
-  delete(id: number): void;
-}
+import { DatabaseSync } from 'node:sqlite';
 
 export const database = new DatabaseSync(':memory:');
 
@@ -49,63 +41,6 @@ CREATE TABLE Announcements (
 
 INSERT INTO Bookings (booking_id, user_id, service_id, booking_datetime, notes) VALUES (1, NULL, NULL, '2025-10-10', 'note');
 `);
-
-interface SqLiteDatabaseOptions<T extends ZodObject> {
-  zodType: T;
-  database: DatabaseSync;
-  getAllStatement: StatementSync;
-  getSingleStatement: StatementSync;
-  updateStatement: StatementSync;
-  insertStatement: StatementSync;
-  deleteStatement: StatementSync;
-}
-
-class SqLiteDatabaseCollection<T extends ZodObject, TObj = z.infer<T>>
-  implements DatabaseCollection<TObj>
-{
-  constructor(options: SqLiteDatabaseOptions<T>) {
-    this.zodType = options.zodType;
-    this.database = database;
-    this.getAllStatement = options.getAllStatement;
-    this.getSingleStatement = options.getSingleStatement;
-    this.updateStatement = options.updateStatement;
-    this.insertStatement = options.insertStatement;
-    this.deleteStatement = options.deleteStatement;
-  }
-
-  private zodType: T;
-  private database: DatabaseSync;
-  private getAllStatement: StatementSync;
-  private getSingleStatement: StatementSync;
-  private updateStatement: StatementSync;
-  private insertStatement: StatementSync;
-  private deleteStatement: StatementSync;
-
-  getAll(): TObj[] {
-    const value = this.getAllStatement.all();
-    console.log(value);
-    return z.array(this.zodType).parse(value);
-  }
-
-  getSingle(id: number): TObj {
-    const value = this.getSingleStatement.get(id);
-    return this.zodType.parse(value);
-  }
-
-  update(entry: TObj): void {
-    // TODO: handle complex data type such as objects.
-    this.updateStatement.run(entry as Record<string, SQLInputValue>);
-  }
-
-  insert(entry: TObj): void {
-    // TODO: handle complex data type such as objects.
-    this.insertStatement.run(entry as Record<string, SQLInputValue>);
-  }
-
-  delete(id: number): void {
-    this.deleteStatement.run(id);
-  }
-}
 
 export const bookings = new SqLiteDatabaseCollection({
   database,
